+++ packages/desktop-client/src/modals/modalsSlice.ts
@@ -265,9 +265,10 @@ export type Modal =
       options: {
         title?: string;
         categoryGroups?: CategoryGroupEntity[];
-        onSelect: (categoryId: string, categoryName: string) => void;
+        onSelect: (categoryId: string | null, categoryName: string) => void;
         month?: string | undefined;
         showHiddenCategories?: boolean;
+        showNoneOption?: boolean;
         closeOnSelect?: boolean;
         clearOnSelect?: boolean;
         onClose?: () => void;